ci(workflows): rename job identifiers and variables for clarity
Rename job identifiers from "python-quality" to "python", "file-linting"
to "lint", and "infrastructure-lint" to "infrastructure" for improved
clarity and simplicity. Change matrix variable from "name" to "type"
to better reflect its purpose, which is to specify the type of files
being linted. This change enhances readability and understanding of
the CI workflow configuration, making it easier to maintain and extend.