ci(ci.yml): update poetry install commands to include all necessary groups
Change the poetry install command to use the `--with` option for better
clarity and to ensure all necessary dependency groups are installed.
In the first job, the `dev` and `types` groups are explicitly included,
ensuring development and type-checking dependencies are installed. In
the second job, the `dev`, `test`, and `types` groups are included to
ensure all dependencies required for testing and development are
available. This change improves the reliability and consistency of the
CI workflow by ensuring all necessary dependencies are installed.