ci(ci.yml): remove fail_ci_if_error option from Codecov action
The `fail_ci_if_error` option is removed from the Codecov action
configuration. This option was set to false, which means that any
errors in the Codecov action would not cause the CI to fail. By
removing this option, the default behavior is used, which is to
fail the CI if there is an error in the Codecov action. This change
ensures that any issues with code coverage reporting are caught
and addressed promptly, maintaining the integrity of the CI process.