ci(docker.yml): remove branch-specific trigger for Docker workflow
Remove the branch-specific trigger for the Docker Build & Deploy
workflow to ensure that the workflow runs on all branches when a
tag is pushed. This change allows for more flexibility in the
workflow execution, ensuring that Docker images are built and
deployed for any branch as long as a version tag is present.