ci(docker.yml): allow Trivy scan to continue on error and adjust exit code
Add `continue-on-error: true` to Trivy scan steps to ensure the workflow
continues even if vulnerabilities are found. Modify `trivy_flags` to
include `--exit-code 0` to prevent the scan from failing the job on
detection of HIGH or CRITICAL vulnerabilities.

These changes ensure that the CI pipeline does not break due to
vulnerability findings, allowing for further analysis and remediation
without halting the development process.