feat(event_handler): integrate onboarding service for new guilds
This commit adds the GuildOnboardingService to the EventHandler class, initializing onboarding for new guilds when they join. The onboarding process is triggered in the on_guild_join method, enhancing the setup experience for new communities. A TODO comment has also been added to define a data expiration policy for guilds in the future.