chore(dependencies): update package versions in uv.lock
This commit updates the versions of several packages in the `uv.lock` file, including `aioconsole` (0.8.1 to 0.8.2), `aiofiles` (24.1.0 to 25.1.0), `aiohttp` (3.12.15 to 3.13.0), `alembic` (1.16.5 to 1.17.0), `yamlfix` (1.18.0 to 1.19.0), and `yarl` (1.20.1 to 1.22.0). The changes include updated source distribution URLs and hashes for the new versions, ensuring compatibility and access to the latest features and fixes.