chore(dependencies): update pre-commit hook versions
This commit updates the versions of several pre-commit hooks in the `.pre-commit-config.yaml` file, including `actionlint` (v1.7.7 to v1.7.8), `pyupgrade` (v3.20.0 to v3.21.0), `add-trailing-comma` (v3.2.0 to v4.0.0), `ruff` (v0.13.2 to v0.14.0), and `commitlint` (v9.22.0 to v9.23.0). These updates ensure compatibility with the latest features and improvements in the respective tools.